[FEAT] #24 - 약추가-원페이지 로직 변경
